+++ EPPZGeometry/EPPZGeometry_PolygonSource.cs
@@ -19,8 +19,8 @@ namespace EPPZGeometry
 
 		void Awake()
 		{
-			// Construct a polygon model from transforms.
-			polygon = Polygon.PolygonWithSource(this);
+			// Construct a polygon model from transforms (if not created by a root polygon already).
+			if (polygon == null) polygon = Polygon.PolygonWithSource (this);
 		}
 
 		void Update()