+++ Test scenes/Controllers/TestScene_02_Controller.cs
@@ -17,11 +17,11 @@ public class TestScene_02_Controller : MonoBehaviour
 	public Material polygonMaterial;
 	public Material intersectingMaterial;
 
-	public EPPZGeometry_PolygonSource starSource;
+	public EPPZGeometry_PolygonSource polygonSource;
 	public GameObject pointSource;
-	public EPPZDebug_PolygonDebugRenderer starRenderer;
+	public EPPZDebug_PolygonDebugRenderer polygonRenderer;
 
-	private Polygon star { get { return starSource.polygon; } }
+	private Polygon polygon { get { return polygonSource.polygon; } }
 	private Vector2 point  { get { return pointSource.transform.position.xy(); } }
 
 
@@ -32,14 +32,14 @@ public class TestScene_02_Controller : MonoBehaviour
 
 	bool PointContainmentTest()
 	{
-		return star.PermiterContainsPoint(point, accuracy, Segment.ContainmentMethod.Precise);
+		return polygon.PermiterContainsPoint(point, accuracy, Segment.ContainmentMethod.Precise);
 	}
 
 	void RenderTestResult(bool testResult)
 	{
 		// Set corresponding materials.
 		Material renderMaterial = (testResult) ? intersectingMaterial : polygonMaterial;
-		starRenderer.lineMaterial = renderMaterial;
+		polygonRenderer.lineMaterial = renderMaterial;
 		pointSource.renderer.material = renderMaterial;
 	}
 }