+++ Test scenes/Controllers/TestScene_052_Controller.cs
@@ -4,7 +4,7 @@ using EPPZGeometry;
 using EPPZDebug;
 
 
-public class TestScene_05_Controller : MonoBehaviour
+public class TestScene_052_Controller : MonoBehaviour
 {
 
 
@@ -14,12 +14,12 @@ public class TestScene_05_Controller : MonoBehaviour
 	public Material polygonMaterial;
 	public Material intersectingMaterial;
 
-	public EPPZGeometry_PolygonSource starSource;
+	public EPPZGeometry_PolygonSource polygonSource;
 	public EPPZGeometry_PolygonSource squareSource;
-	public EPPZDebug_PolygonDebugRenderer starRenderer;
+	public EPPZDebug_PolygonDebugRenderer polygonRenderer;
 	public EPPZDebug_PolygonDebugRenderer squareRenderer;
 
-	private Polygon star { get { return starSource.polygon; } }
+	private Polygon polygon { get { return polygonSource.polygon; } }
 	private Polygon square { get { return squareSource.polygon; } }
 		
 
@@ -34,26 +34,22 @@ public class TestScene_05_Controller : MonoBehaviour
 		bool pointContainment = true;
 		square.EnumeratePointsRecursive((Vector2 eachPoint) =>
 		{
-			pointContainment &= star.ContainsPoint(eachPoint);
+			pointContainment &= polygon.ContainsPoint(eachPoint);
 		});
 
 		// Segment-Polygon intersecion, Segment endpoint-permiter contaimnent.
 		bool segmentIntersecting = false;
-		bool permiterContainsSegment = false;
 		foreach (Edge eachEdge in square.edges)
 		{
-			permiterContainsSegment |= star.PermiterContainsPoint(eachEdge.a) || star.PermiterContainsPoint(eachEdge.b);
-			segmentIntersecting |= star.IsIntersectingWithSegment(eachEdge);
+			segmentIntersecting |= polygon.IsIntersectingWithSegment(eachEdge);
 		}
 
 		// A polygon contains a foreign polygon, when
 		// + foreign polygon vertices are contained by polygon
 		// + foreign polygon segments are not intersecting with polygon
-		// + foreign polygon vertices are not contained by polygon permiter
 		bool polygonInside = (
 			pointContainment &&
-			segmentIntersecting == false &&
-			permiterContainsSegment == false
+			segmentIntersecting == false
 			);
 
 		return polygonInside;
@@ -63,7 +59,7 @@ public class TestScene_05_Controller : MonoBehaviour
 	{
 		// Set corresponding materials.
 		Material renderMaterial = (testResult) ? intersectingMaterial : polygonMaterial;
-		starRenderer.lineMaterial = renderMaterial;
+		polygonRenderer.lineMaterial = renderMaterial;
 		squareRenderer.lineMaterial = renderMaterial;
 	}
 }